requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-phonenumber-field==6.1.0
django-private-storage==3.0
filelock==3.7.0
idna==3.3
phonenumbers==8.12.48
pkg_resources==0.0.0
platformdirs==2.5.2
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.9.1
urllib3==1.26.9
uWSGI==2.0.20
virtualenv==20.14.1
virtualenv-clone==0.5.7
